requirements.txt
fastapi
uvicorn
gunicorn
motor
pymongo
pydantic
passlib
python-dotenv
httpx
openai
slack-sdk